+++ s4e-web/src/app/utils.ts
@@ -12,7 +12,3 @@ export function activateMatcher(url: UrlSegment[]) {
   }
   return null;
 }
-
-export function hasBeenClickedOutside(source, target): boolean {
-  return !source.nativeElement.contains(target);
-}