+++ s4e-web/src/app/utils/dropdown/events.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { EventsDirective } from './events.directive';
+
+@NgModule({
+  declarations: [EventsDirective],
+  imports: [
+    CommonModule
+  ],
+  exports: [EventsDirective]
+})
+export class EventsModule { }