+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The most recent changes are on top, in each type of changes category.
 ## [Unreleased]
 
 ### Added
-
+- Create events directive [#592](https://github.com/cyfronet-fid/sat4envi/issues/592)
 
 ### Changed